pyproject.toml
[tool.poetry]
name = "GBKviz"
version = "0.0.0"
description = "Genbank data visualization tool"
authors = ["moshi"]

[tool.poetry.dependencies]
python = "^3.8"
biopython = "^1.79"
reportlab = "^3.6.2"
streamlit = "^1.1.0"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
flake8 = "^4.0.1"
pydocstyle = "^6.1.1"
ipykernel = "^6.5.0"
ipython = "^7.29.0"

[tool.poetry.scripts]
gbkviz_webapp = "gbkviz.scripts.launch_st_gbkviz:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
